final cleanup for integration test pr
