+++ components/sync15/src/lib.rs
@@ -23,7 +23,6 @@ mod sync_multiple;
 pub mod telemetry;
 mod token;
 mod util;
-mod engine;
 
 // Re-export some of the types callers are likely to want for convenience.
 pub use crate::bso_record::{BsoRecord, CleartextBso, EncryptedBso, EncryptedPayload, Payload};
@@ -44,5 +43,3 @@ pub use crate::sync_multiple::{
     sync_multiple, sync_multiple_with_command_processor, MemoryCachedState, SyncRequestInfo,
 };
 pub use crate::util::ServerTimestamp;
-// REMEMBER: this exposes the objects to be imported/used by files in other components.
-pub use crate::engine::{TestRecord, TestStore, SyncMultipleStorage, SyncMultipleEngine};
