--- components/sync15/src/sync_multiple_unit_tests.rs
@@ -1,23 +1,0 @@
-use crate::sync_multiple::{MemoryCachedState};
-use crate::client::{Sync15StorageClientInit};
-use crate::key_bundle::KeyBundle;
-
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn basic_test() {
-        let mut mem_cached_state = MemoryCachedState {
-
-        }
-    }
-
-}
-
-// Dummy; empty.
-pub struct MockStore {
-}
-impl Store for MockStore {
-}
