+++ testing/sync-test/Cargo.toml
@@ -23,3 +23,4 @@ lazy_static = "1.4.0"
 structopt = "0.3.11"
 serde = "1.0.104"
 serde_derive = "1.0.104"
+serde_json = "1.0.50"
